2998-data-table-query-tool: Initial commit of sales data table
