+++ src/components/data-viz/QueryToolTable/QueryToolTablePeriod/index.js
@@ -1,0 +1,1 @@
+export default from './QueryToolTablePeriod'
