+++ src/components/data-viz/QueryToolTable/QueryToolTableSales/index.js
@@ -1,0 +1,1 @@
+export default from './QueryToolTableSales'
